Fix build errors on Windows and Fedora for merged PR HTCONDOR-149

+++ src/condor_submit.V6/submit.cpp
@@ -2657,7 +2657,7 @@ bool get_oauth_service_requests(std::string & service_requests) {
 			service_requests += " ";
 		service_requests += str;
 		std::string keys[] = { "handle", "scopes", "audience" };
-		for (int i = 0; i < (sizeof(keys)/sizeof(keys[0])); i++ ) {
+		for (size_t i = 0; i < (sizeof(keys)/sizeof(keys[0])); i++ ) {
 			str = "";
 			request->LookupString(keys[i], str);
 			if (str != "") {
